backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client>=1.8.0
fastembed==0.2.0
openai-agents==0.2.9
python-dotenv==1.0.0
pydantic==2.5.0